requirements.txt
flask==2.3.3
gspread==5.12.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0